requirements.txt
# Bot Telegram et serveur web
python-telegram-bot[webhooks]==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Client Notion
notion-client==2.2.1

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Tests et qualité de code
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.1.6

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.0

